rewrite: cli rewrite
