+++ anime_downloader/__version__.py
@@ -1,1 +1,1 @@
-__version__ = '3.6.3'
+__version__ = '4.0.0.r1'
