+++ anime_downloader/config.py
@@ -15,7 +15,6 @@ DEFAULT_CONFIG = {
         'quality': '720p',
         'fallback_qualities': ['720p', '480p', '360p'],
         'force_download': False,
-        'log_level': 'INFO',
         'file_format': '{anime_title}/{anime_title}_{ep_no}',
         'provider': '9anime',
         'external_downloader': '',
