+++ docs/api/anime.rst
@@ -3,7 +3,7 @@ Base classes
 
 .. automodule:: anime_downloader.sites.anime
 
-.. autoclass:: anime_downloader.sites.anime.BaseAnime
-   :members: search, get_data, _scarpe_episodes, _scrape_metadata
+.. autoclass:: anime_downloader.sites.anime.Anime
+   :members: search, get_data, _scrape_episodes, _scrape_metadata
 
 .. autoclass:: anime_downloader.sites.anime.SearchResult
